fixes test fails and cleans up test set up
 - impr: shared/installed holds word2vec and sentence-transformer downloads
 - impr: replaces virtual_env w default-python venv
 - impr: 'make format' also adds licenses
