+++ mentor_classifier/Makefile
@@ -5,8 +5,8 @@ $(VENV):
 	$(MAKE) $(VENV)-update
 
 .PHONY: $(VENV)-update
-$(VENV)-update: virtualenv-installed
-	[ -d $(VENV) ] || virtualenv -p python3.8 $(VENV)
+$(VENV)-update:
+	[ -d $(VENV) ] || python3.8 -m venv $(VENV)
 	$(VENV)/bin/pip install --upgrade pip
 	$(VENV)/bin/pip install -r ./requirements.test.txt
 	$(VENV)/bin/python3 -m nltk.downloader averaged_perceptron_tagger
@@ -55,7 +55,8 @@ license:
 PHONY: test
 test: $(VENV)
 	rm -rf tests/fixtures/data_out/*/*
-	cd $(ROOT)/word2vec/ && $(MAKE) word2vec.bin
+	cd $(ROOT)/shared && $(MAKE) installed/word2vec.bin
+	cd $(ROOT)/shared && $(MAKE) installed/sentence-transformer
 	. $(VENV)/bin/activate \
 		&& coverage run \
 			--omit="$(PWD)/tests $(VENV)" \
@@ -83,6 +84,3 @@ test-types: $(VENV)
 .PHONY: update-deps
 update-deps: $(VENV)
 	. $(VENV)/bin/activate && pip-upgrade requirements*
-
-virtualenv-installed:
-	$(ROOT)/tools/virtualenv_ensure_installed.sh
