+++ mentor_classifier/mentor_classifier/arch/lr_transformer/embeddings.py
@@ -15,7 +15,6 @@ class TransformerEmbeddings:
         self.transformer: SentenceTransformer = find_or_load_sentence_transformer(
             path.join(shared_root, "sentence-transformer")
         )
-        # self.transformer = SentenceTransformer('distilbert-base-nli-mean-tokens', device=0)  # This is to use GPU
 
     def get_embeddings(self, data: List[str]):
         embeddings = self.transformer.encode(data, show_progress_bar=True)
