+++ mentor_classifier/tests/conftest.py
@@ -9,10 +9,5 @@ import pytest
 
 
 @pytest.fixture(scope="module", autouse=True)
-def word2vec() -> str:
-    word2vec_path = path.abspath(path.join("..", "word2vec", "word2vec.bin"))
-    if not path.exists(word2vec_path):
-        raise Exception(
-            "missing required word2vec.bin. Try run `make clean word2vec.bin` from <PROJECT_ROOT>/word2vec folder"
-        )
-    return word2vec_path
+def shared_root() -> str:
+    return path.abspath(path.join("..", "shared", "installed"))
