+++ mentor_classifier/tests/test_mentor.py
@@ -11,6 +11,7 @@ import pytest
 from mentor_classifier.mentor import Mentor
 from .helpers import fixture_path
 
+
 @responses.activate
 @pytest.mark.parametrize(
     "mentor_id,expected_data",
@@ -35,7 +36,9 @@ from .helpers import fixture_path
                         ]
                     ],
                     "_INTRO_": [["A5", "Hi I'm Clint", []]],
-                    "_OFF_TOPIC_": [["A6", "Ask me something else", []]],
+                    "_OFF_TOPIC_": [
+                        ["A6", "Ask me something else", ["/clint/offtopic.mp4"]]
+                    ],
                 },
                 "questions_by_id": {
                     "Q1": {
