+++ mentor_classifier_api/mypy.ini
@@ -16,7 +16,7 @@ ignore_missing_imports = True
 [mypy-mentor_classifier.*]
 ignore_missing_imports = True
 
-[mypy-sentence_transformers.*]
+[mypy-sentence-transformers.*]
 ignore_missing_imports = True
 
 [mypy-mentor_classifier.svm.*]
