+++ shared/utils.py
@@ -4,29 +4,25 @@
 #
 # The full terms of this copyright and license should always be found in the root directory of this software deliverable as "license.txt" and if these terms are not found with this software, please contact the USC Stevens Center for the full license.
 #
-from os import path
-
-import json
-import pytest
-import responses
-
-from mentor_classifier import ClassifierFactory, ARCH_DEFAULT
-from .helpers import fixture_path
+import os
+import requests
+from tqdm import tqdm
 
 
-@pytest.fixture(scope="module")
-def data_root() -> str:
-    return fixture_path("data_out")
-
-
-@responses.activate
-@pytest.mark.parametrize("mentor_id", [("clint")])
-def test_trains_and_outputs_models(data_root: str, shared_root: str, mentor_id: str):
-    with open(fixture_path("graphql/{}.json".format(mentor_id))) as f:
-        data = json.load(f)
-    responses.add(responses.POST, "http://graphql/graphql", json=data, status=200)
-    result = ClassifierFactory().new_training(mentor_id, shared_root, data_root).train()
-    print(result)
-    assert result.model_path == path.join(data_root, mentor_id, ARCH_DEFAULT)
-    assert path.exists(path.join(result.model_path, "model.pkl"))
-    assert path.exists(path.join(result.model_path, "w2v.txt"))
+def download(url: str, to_path: str):
+    print("downloading")
+    print(f"\tfrom {url}")
+    print(f"\tto {to_path}")
+    r = requests.get(url, stream=True)
+    total_size_in_bytes = int(r.headers.get("content-length", 0))
+    block_size = 1024
+    progress_bar = tqdm(total=total_size_in_bytes, unit="iB", unit_scale=True)
+    os.makedirs(os.path.dirname(to_path), exist_ok=True)
+    with open(to_path, "wb") as f:
+        for chunk in r.iter_content(chunk_size=block_size):
+            if chunk:
+                progress_bar.update(len(chunk))
+                f.write(chunk)
+    progress_bar.close()
+    if total_size_in_bytes != 0 and progress_bar.n != total_size_in_bytes:
+        raise Exception("failed to fully download")
