--- word2vec/sentence-transformer/requirements.txt
@@ -1,2 +1,0 @@
-requests==2.25.0
-tqdm==4.52.0
